+++ kitty/options/definition.py
@@ -330,6 +330,12 @@ cursor shape to :code:`beam` at shell prompts. You can avoid this by setting
 '''
     )
 
+opt('cursor_shape_unfocused', 'hollow', option_type='to_cursor_unfocused_shape', ctype='int', long_text='''
+Defines the text cursor shape when the OS window is not focused. The unfocused
+cursor shape can be one of :code:`block`, :code:`beam`, :code:`underline`,
+:code:`hollow`.
+''')
+
 opt('cursor_beam_thickness', '1.5',
     option_type='positive_float', ctype='float',
     long_text='The thickness of the beam cursor (in pts).'
@@ -357,14 +363,6 @@ inactivity. Set to zero to never stop blinking.
 '''
     )
 
-opt('cursor_unfocused_shape', 'hollow',
-    option_type='to_cursor_unfocused_shape', ctype='int',
-    long_text='''
-Defines the cursor shape when the terminal window is not focused. The unfocused
-cursor shape can be one of :code:`block`, :code:`beam`, :code:`underline`,
-:code:`hollow`.
-'''
-    )
 egr()  # }}}