+++ kitty/lineops.h
@@ -89,8 +89,8 @@ void line_set_char(Line *, unsigned int , uint32_t , unsigned int , Cursor *, hy
 void line_right_shift(Line *, unsigned int , unsigned int );
 void line_add_combining_char(CPUCell *, GPUCell *, uint32_t , unsigned int );
 index_type line_url_start_at(Line *self, index_type x);
-index_type line_url_end_at(Line *self, index_type x, bool, char_type, bool);
-bool line_startswith_url_chars(Line*);
+index_type line_url_end_at(Line *self, index_type x, bool, char_type, bool, bool, index_type);
+bool line_startswith_url_chars(Line*, bool);
 bool line_as_ansi(Line *self, ANSIBuf *output, const GPUCell**, index_type start_at, index_type stop_before, char_type prefix_char) __attribute__((nonnull));
 unsigned int line_length(Line *self);
 size_t cell_as_unicode(CPUCell *cell, bool include_cc, Py_UCS4 *buf, char_type);