+++ docs/changelog.rst
@@ -53,7 +53,7 @@ Detailed list of changes
 0.36.0 [future]
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-- A new option, :opt:`window_logo_scale` to specify how window logo are scaled with respect to the size of the window containing the logo (:pull:`7534`)
+- A new option, :opt:`window_logo_scale` to specify how window logos are scaled with respect to the size of the window containing the logo (:pull:`7534`)
 
 - A new option, :opt:`cursor_shape_unfocused` to specify the shape of the text cursor in unfocused OS windows (:pull:`7544`)